Remove dependency on io-console gem
`io/console` part of Ruby and coolline itself works without the gem. As
it turns out, this gem dependency is actually unnecessary and is
preventing pry-coolline from building under Ruby 2.0 (as the io-console
gem has native extensions that fail to configure). Remove the gem
dependency and instead use the `io/console` that already ships with Ruby
1.9.x to also support Ruby 2.0.0

Signed-off-by: David Celis <me@davidcel.is>